Fix non-functional MTP.
This requires a minor patch to the kernel:

  https://github.com/ianmacd/c2s/commit/a1962f1c17c9b928cee9ec18df530aa440b5c71c

and a kernel config change to avoid a reproducible spontaneous reboot
after 220 seconds of uptime in TWRP:

  https://github.com/ianmacd/c2s/commit/56f328d8c56bd795ca222a07460759d884be251e

The default TWRP init.rc is now used in favour of Samsung's recovery
init.rc, and a custom USB configuration is applied.

Additionally, MTP no longer needs to be disabled to allow adb to work.